+++ index.d.ts
@@ -10,7 +10,11 @@ export {
 export {default as iterate} from './iterate';
 export {default as maxValue} from './maxValue';
 export {default as minValue} from './minValue';
-export {default as remove} from './remove';
+export {
+  default as remove,
+  removeIfExists,
+  removeOrThrowIfNotExists,
+} from './remove';
 export {default as reverseIterate} from './reverseIterate';
 export * from './actions';
 export * from './types';